xml/parsing/manakai/attrs-2.dat

#data
<!DOCTYPE a[
<!ATTLIST a
b CDATA "c"
>
]>
<a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b CDATA () EXPLICIT "c"
| <{} a>
|   b="c"

#data
<!DOCTYPE a[
<!ATTLIST a
b CDATA "c"
d ENTITIES "e"
>
]>
<a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b CDATA () EXPLICIT "c"
|     d ENTITIES () EXPLICIT "e"
| <{} a>
|   b="c"
|   d="e"

#data
<!DOCTYPE a[
<!ATTLIST a
b CDATA #FIXED "c"
>
]>
<a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b CDATA () FIXED "c"
| <{} a>
|   b="c"

#data
<!DOCTYPE a[
<!ATTLIST a
b CDATA "c"
>
]>
<a b="d"/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b CDATA () EXPLICIT "c"
| <{} a>
|   b="d"

#data
<!DOCTYPE a[
<!ATTLIST a
xmlns CDATA "c"
>
]>
<a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     xmlns CDATA () EXPLICIT "c"
| <{c} a>
|   xmlns xmlns="c"

#data
<!DOCTYPE a[
<!ATTLIST a
xmlns:b CDATA "c"
>
]>
<a b:e="f"/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     xmlns:b CDATA () EXPLICIT "c"
| <{} a>
|   xmlns b="c"
|   {c} e="f"

#data
<!DOCTYPE a[
<!ATTLIST a
xmlns:b CDATA "c"
>
]>
<a xmlns:b="d" b:e="f"/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     xmlns:b CDATA () EXPLICIT "c"
| <{} a>
|   xmlns b="d"
|   {d} e="f"

#data
<!DOCTYPE a[
<!ATTLIST a
b:c CDATA "d"
xmlns:b CDATA "x"
>
]>
<a e:c="f" xmlns:e="x"/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b:c CDATA () EXPLICIT "d"
|     xmlns:b CDATA () EXPLICIT "x"
| <{} a>
|   b:c="d"
|   xmlns b="x"
|   xmlns e="x"
|   {x} c="f"

#data escaped
<!DOCTYPE a[
<!ATTLIST a
b CDATA " c\u0009d\u000a"
>
]>
<a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b CDATA () EXPLICIT " c d "
| <{} a>
|   b=" c d "

#data escaped
<!DOCTYPE a[
<!ATTLIST a
b NMTOKEN " c  \u0009d\u000a "
>
]>
<a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b NMTOKEN () EXPLICIT "c d"
| <{} a>
|   b="c d"

#data
<!DOCTYPE a[
<!ATTLIST a
b IDREF "c &#x9;d &#xa; "
>
]>
<a/>
#errors
1;1;no XML decl;;;s
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b IDREF () EXPLICIT "c \u0009d \u000a"
| <{} a>
|   b="c \u0009d \u000a"

#data escaped
<!DOCTYPE a[
<!ATTLIST a
b NMTOKENS #IMPLIED
>
]>
<a b="c &#x9;d  e\u000a"/>
#errors
1;1;no XML decl;;;s
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b NMTOKENS () IMPLIED ""
| <{} a>
|   b="c \u0009d e"

#data escaped
<!DOCTYPE a[
<!ATTLIST a
xmlns NMTOKEN " x  y \u0009"
>
]>
<a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     xmlns NMTOKEN () EXPLICIT "x y"
| <{x y} a>
|   xmlns xmlns="x y"

#data
<!DOCTYPE a[
<!ATTLIST a
b CDATA "c"
>
]>
<z><a/></z>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b CDATA () EXPLICIT "c"
| <{} z>
|   <{} a>
|     b="c"

#data
<!DOCTYPE a[
<!ATTLIST a
b CDATA "c"
d ENTITIES "e"
>
]>
<z><a/></z>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b CDATA () EXPLICIT "c"
|     d ENTITIES () EXPLICIT "e"
| <{} z>
|   <{} a>
|     b="c"
|     d="e"

#data
<!DOCTYPE a[
<!ATTLIST a
b CDATA #FIXED "c"
>
]>
<z><a/></z>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b CDATA () FIXED "c"
| <{} z>
|   <{} a>
|     b="c"

#data
<!DOCTYPE a[
<!ATTLIST a
b CDATA "c"
>
]>
<z><a b="d"/></z>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b CDATA () EXPLICIT "c"
| <{} z>
|   <{} a>
|     b="d"

#data
<!DOCTYPE a[
<!ATTLIST a
xmlns CDATA "c"
>
]>
<z><a/></z>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     xmlns CDATA () EXPLICIT "c"
| <{} z>
|   <{c} a>
|     xmlns xmlns="c"

#data
<!DOCTYPE a[
<!ATTLIST a
xmlns:b CDATA "c"
>
]>
<z><a b:e="f"/></z>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     xmlns:b CDATA () EXPLICIT "c"
| <{} z>
|   <{} a>
|     xmlns b="c"
|     {c} e="f"

#data
<!DOCTYPE a[
<!ATTLIST a
xmlns:b CDATA "c"
>
]>
<z><a xmlns:b="d" b:e="f"/></z>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     xmlns:b CDATA () EXPLICIT "c"
| <{} z>
|   <{} a>
|     xmlns b="d"
|     {d} e="f"

#data
<!DOCTYPE a[
<!ATTLIST a
b:c CDATA "d"
xmlns:b CDATA "x"
>
]>
<z><a e:c="f" xmlns:e="x"/></z>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b:c CDATA () EXPLICIT "d"
|     xmlns:b CDATA () EXPLICIT "x"
| <{} z>
|   <{} a>
|     b:c="d"
|     xmlns b="x"
|     xmlns e="x"
|     {x} c="f"

#data escaped
<!DOCTYPE a[
<!ATTLIST a
b CDATA " c\u0009d\u000a"
>
]>
<z><a/></z>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b CDATA () EXPLICIT " c d "
| <{} z>
|   <{} a>
|     b=" c d "

#data escaped
<!DOCTYPE a[
<!ATTLIST a
b NMTOKEN " c  \u0009d\u000a "
>
]>
<z><a/></z>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b NMTOKEN () EXPLICIT "c d"
| <{} z>
|   <{} a>
|     b="c d"

#data
<!DOCTYPE a[
<!ATTLIST a
b IDREF "c &#x9;d &#xa; "
>
]>
<z><a/></z>
#errors
1;1;no XML decl;;;s
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b IDREF () EXPLICIT "c \u0009d \u000a"
| <{} z>
|   <{} a>
|     b="c \u0009d \u000a"

#data escaped
<!DOCTYPE a[
<!ATTLIST a
b NMTOKENS #IMPLIED
>
]>
<z><a b="c &#x9;d  e\u000a"/></z>
#errors
1;1;no XML decl;;;s
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b NMTOKENS () IMPLIED ""
| <{} z>
|   <{} a>
|     b="c \u0009d e"

#data escaped
<!DOCTYPE a[
<!ATTLIST a
xmlns NMTOKEN " x  y \u0009"
>
]>
<z><a/></z>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     xmlns NMTOKEN () EXPLICIT "x y"
| <{} z>
|   <{x y} a>
|     xmlns xmlns="x y"

#data escaped
<!DOCTYPE a [
  <!ENTITY c " e\u0009d ">
]><a b="&c;"/>
#errors
1;1;no XML decl;;;s
#document escaped
| <!DOCTYPE a>
| <{} a>
|   b=" e d "

#data escaped
<!DOCTYPE a [
  <!ENTITY c " e\u0009d ">
  <!ENTITY d "&c;">
]><a b="&d;"/>
#errors
1;1;no XML decl;;;s
#document escaped
| <!DOCTYPE a>
| <{} a>
|   b=" e d "

#data escaped
<!DOCTYPE a [
  <!ENTITY c " e\u000Ad ">
  <!ENTITY d "&c;">
]><a b="&d;"/>
#errors
1;1;no XML decl;;;s
#document escaped
| <!DOCTYPE a>
| <{} a>
|   b=" e d "

#data escaped
<!DOCTYPE a [
  <!ENTITY c " e\u000Dd ">
  <!ENTITY d "&c;">
]><a b="&d;"/>
#errors
1;1;no XML decl;;;s
#document escaped
| <!DOCTYPE a>
| <{} a>
|   b=" e d "

#data escaped
<!DOCTYPE a [
  <!ENTITY c " e&#x9;d ">
  <!ENTITY d "&c;">
]><a b="&d;"/>
#errors
1;1;no XML decl;;;s
#document escaped
| <!DOCTYPE a>
| <{} a>
|   b=" e d "

#data escaped
<!DOCTYPE a [
  <!ENTITY c " e\u0009d ">
  <!ENTITY d "&c;\u000A">
]><a b="&d;"/>
#errors
1;1;no XML decl;;;s
#document escaped
| <!DOCTYPE a>
| <{} a>
|   b=" e d  "

#data escaped
<!DOCTYPE a [
  <!ENTITY c " e\u0009d ">
  <!ENTITY d "&c;&#10;">
]><a b="&d;"/>
#errors
1;1;no XML decl;;;s
#document escaped
| <!DOCTYPE a>
| <{} a>
|   b=" e d  "

#data escaped
<!DOCTYPE a [
  <!ENTITY c " e\u000Dd ">
  <!ENTITY d "&c;\u000A">
]><a b="&d;"/>
#errors
1;1;no XML decl;;;s
#document escaped
| <!DOCTYPE a>
| <{} a>
|   b=" e d  "

#data escaped
<!DOCTYPE a [
  <!ENTITY c " e\u0009d ">
  <!ENTITY d "&c;&#10;">
]><a b="&d;"/>
#errors
1;1;no XML decl;;;s
#document escaped
| <!DOCTYPE a>
| <{} a>
|   b=" e d  "

#data escaped
<a b="&#13;"/>
#errors
1;1;no XML decl;;;s
#document escaped
| <{} a>
|   b="\u000D"

#data escaped
<a b="&#10;"/>
#errors
1;1;no XML decl;;;s
#document escaped
| <{} a>
|   b="\u000A"

#data escaped
<a b="&#x9;"/>
#errors
1;1;no XML decl;;;s
#document escaped
| <{} a>
|   b="\u0009"

#data escaped
<a b="\u000D"/>
#errors
1;1;no XML decl;;;s
#document escaped
| <{} a>
|   b=" "

#data escaped
<!DOCTYPE a [
  <!ENTITY c " e&#x9;d ">
  <!ENTITY d "&c;">
  <!ATTLIST a b ID #IMPLIED>
]><a b="&d;"/>
#errors
1;1;no XML decl;;;s
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b ID () IMPLIED ""
| <{} a>
|   b="e d"
